Rating is a float and get service with id returns service with images

+++ TripPlanner.Application/Rating/Commands/AddRating/AddRatingCommand.cs
@@ -11,6 +11,6 @@ namespace TripPlanner.Application.Rating.Commands.AddRating
     {
         public int? ServiceId { get; set; }
         public string? UserId { get; set; }
-        public int Rating { get; set; } = 0;
+        public float Rating { get; set; } = 0;
     }
 }
